pyproject.toml
[project]
name = "volt"
description = "Extensible static site generator"
version = "0.0dev0"
authors = [
    {name = "Wibowo Arindrarto", email = "contact@arindrarto.dev"},
]
requires-python = "==3.12.*"
dependencies = [
    "blinker==1.*",
    "click==8.*",
    "jinja2==3.*",
    "markdown2==2.*",
    "marshmallow==3.*",
    "pendulum==3.*",
    "pygments==2.*",
    "python-slugify==8.*",
    "pyyaml==6.*",
    "rich==13.*",
    "structlog==24.*",
    "tomlkit==0.*",
    "urllib3==2.*",
    "watchdog==2.2.1",
]
readme = "README.md"
keywords = ["static website"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Utilities",
]

[project.urls]
changelog = "https://github.com/bow/volt/blob/master/CHANGELOG.rst"
repository = "https://github.com/bow/volt"

[project.scripts]
volt = "volt.cli:main"

[dependency-groups]
dev = [
    "bandit==1.*",
    "insipid-sphinx-theme==0.*",
    "mypy==1.*",
    "mypy-extensions==1.*",
    "pytest==8.*",
    "pytest-cov==5.*",
    "pytest-mock==3.*",
    "pytest-structlog==1.*",
    "pytest-sugar==1.*",
    "radon==6.*",
    "requests==2.*",
    "safety==3.*",
    "sphinx==8.*",
    "toml==0.*",
    "types-python-slugify==8.*",
    "types-pyyaml==6.*",
    "types-requests==2.*",
    "wheel==0.*",
]

[tool.uv]
package = true

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = [
    "E203", "E731",
    "F403", "F405",
    "UP007", "UP045",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
show_error_context = true
show_column_numbers = true
show_error_codes = true
pretty = true
warn_redundant_casts = true

[[tool.mypy.overrides]]
module = "volt.*"
disallow_untyped_calls = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
allow_untyped_globals = false
strict_equality = true

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'
exclude = '/(\.eggs|\.git|\.mypy_cache|\.tox|\.venv|_build|build|dist)/'

[tool.coverage.run]
branch = true
source = ["src/volt"]

[tool.coverage.report]
exclude_lines = [
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
omit = ["tests/*"]

[tool.distutils.bdist_wheel]
universal = true
